Update the-changelog-316.md
Fixed a bunch of 'unintelligibles'. 

Could not for the life of me catch what was said at **00:15:20.13**... Maybe someone else can hear that in a way that makes sense. I'll definitely be making other PRs as well for other 'unintelligibles' in other episodes; would've done so sooner, but didn't even know such an option (to help fix the transcripts) was available!