Fix unintelligible: go vendoring
This was spoken pretty fast but considering the context I think it's
clear what the words are and it's probably a reference to this having
been a topic on previous episodes.

The change above it for "that stuff's" is grammatical, but it also
sounds more like what's being said.  I don't mind if that change doesn't
get merged though.

In Go Time 105.