Fix unintelligible: dynacism
My best guess here is dynacism, of which I was unable to find any normal
occurrences online, other than that it's apparently a common misspelling
(or in this case misspeaking) of dynamism.  My text editor also keeps
trying to correct it accordingly.  It has several meanings, of which the
latter two may be relevant:

1. Any of various theories or philosophical systems that explain the
   universe in terms of force or energy.
2. A process or mechanism responsible for the development or motion of a
   system.
3. Continuous change, activity, or progress; vigour.

Especially the last definition would make sense in this context, so I'd
say this is most likely the word that was being used here.

In Go Time 105.