ship-it-37: various minor corrections in transcripts
Mostly fixes to things that have names which are easily misheard (e.g.
"Nix packages" -> nixpkgs, which is pronounced the same).

Also wrapped all shell commands etc. that we pronounced out loud in
backticks to get monospaced formatting. I verified in other Changelog
podcast transcripts that this renders as intended.